+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/analytics/MissingResultFeedbackEvent.kt
@@ -29,6 +29,14 @@ public class MissingResultFeedbackEvent @JvmOverloads public constructor(
      * Unique ID for identifying several related Mapbox events per session.
      */
     public val sessionId: String? = null,
+
+    /**
+     * Unique feedback ID. If the passed value is null, the id will be generated inside the Search SDK.
+     * Normally SDK users don't have to provide it explicitly.
+     * One of the cases, when users should provide their own id,
+     * is when they want to have associated events in different analytics systems.
+     */
+    public val feedbackId: String? = null,
 ) {
 
     /**
@@ -40,12 +48,14 @@ public class MissingResultFeedbackEvent @JvmOverloads public constructor(
         text: String? = this.text,
         screenshot: Bitmap? = this.screenshot,
         sessionId: String? = this.sessionId,
+        feedbackId: String? = this.feedbackId,
     ): MissingResultFeedbackEvent {
         return MissingResultFeedbackEvent(
             responseInfo = responseInfo,
             text = text,
             screenshot = screenshot,
             sessionId = sessionId,
+            feedbackId = feedbackId,
         )
     }
 
@@ -62,6 +72,7 @@ public class MissingResultFeedbackEvent @JvmOverloads public constructor(
         if (text != other.text) return false
         if (screenshot != other.screenshot) return false
         if (sessionId != other.sessionId) return false
+        if (feedbackId != other.feedbackId) return false
 
         return true
     }
@@ -74,6 +85,7 @@ public class MissingResultFeedbackEvent @JvmOverloads public constructor(
         result = 31 * result + (text?.hashCode() ?: 0)
         result = 31 * result + (screenshot?.hashCode() ?: 0)
         result = 31 * result + (sessionId?.hashCode() ?: 0)
+        result = 31 * result + (feedbackId?.hashCode() ?: 0)
         return result
     }
 
@@ -85,7 +97,8 @@ public class MissingResultFeedbackEvent @JvmOverloads public constructor(
                 "responseInfo=$responseInfo, " +
                 "text=$text, " +
                 "screenshot=$screenshot, " +
-                "sessionId=$sessionId" +
+                "sessionId=$sessionId, " +
+                "feedbackId=$feedbackId" +
                 ")"
     }
 }